moved to generic validator class
